feat(src/components/CardDescription/index.tsx) fizemos a logica de divisao das colunas @fernandesmelo @Lip 29bm
